retrofit sustained to singleton
