+++ app/src/main/java/com/electrolux/test/App.kt
@@ -1,0 +1,35 @@
+package com.electrolux.test
+
+import android.app.Application
+import com.electrolux.test.api.IFlickrRequest
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+class App: Application() {
+
+    lateinit var flickrApi: IFlickrRequest
+
+    override fun onCreate() {
+        super.onCreate()
+
+        configureRetrofit()
+    }
+
+    private fun configureRetrofit() {
+        val logging = HttpLoggingInterceptor()
+        logging.setLevel(HttpLoggingInterceptor.Level.BODY)
+        val client: OkHttpClient = OkHttpClient.Builder()
+            .addInterceptor(logging)
+            .build()
+
+        val retrofit = Retrofit.Builder()
+            .client(client)
+            .addConverterFactory(GsonConverterFactory.create())
+            .baseUrl("https://api.flickr.com")
+            .build()
+
+        flickrApi = retrofit.create(IFlickrRequest::class.java)
+    }
+}
